logo fixing

+++ lib/powerpoint/util.rb
@@ -25,7 +25,7 @@ module Powerpoint
     def copy_media(extract_path, image_path)
       image_name = File.basename(image_path)
       dest_path = "#{extract_path}/ppt/media/#{image_name}"
-      FileUtils.copy_file(image_path, dest_path)
+      FileUtils.copy_file(image_path, dest_path) unless File.exist?(dest_path)
     end
 
     def merge_variables(b, variables)
